refactor(runbotics-scheduler): :recycle: Changes after CR
