+++ runbotics/runbotics-scheduler/src/mail-trigger/mail-trigger.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { ScheduleModule } from '@nestjs/schedule';
+import { DatabaseModule } from '#/database/database.module';
+import { QueueModule } from '#/queue/queue.module';
+import { MailTriggerService } from './mail-trigger.service';
+
+@Module({
+    imports: [
+        DatabaseModule,
+        QueueModule,
+        ScheduleModule.forRoot(),
+    ],
+    providers: [MailTriggerService],
+    exports: [MailTriggerService],
+})
+export class MailTriggerModule {}
