+++ runbotics/runbotics-scheduler/src/websocket/websocket.module.ts
@@ -4,7 +4,7 @@ import { QueueModule } from '#/queue/queue.module';
 import { MicrosoftModule } from '#/microsoft';
 import { DatabaseModule } from '../database/database.module';
 import { AuthModule } from '../auth/auth.module';
-import { MailModule } from '../mail/mail.module';
+import { MailTriggerModule } from '../mail-trigger/mail-trigger.module';
 import { BotProcessService } from './bot/process-launch/bot-process-instance.service';
 import { BotProcessEventService } from './bot/process-launch/bot-process-instance-event.service';
 import { BotLogService } from './bot/bot-log.service';
@@ -13,7 +13,7 @@ import { WebsocketService } from './websocket.service';
 import { UiGateway } from './ui/ui.gateway';
 import { BotLifecycleService } from './bot/bot-lifecycle.service';
 import { Logger } from 'src/utils/logger';
-import { NestMailerModule } from '#/nest-mailer/nest-mailer.module';
+import { MailModule } from '#/mail/mail.module';
 import { ProcessModule } from '#/database/process/process.module';
 
 @Global()
@@ -21,10 +21,10 @@ import { ProcessModule } from '#/database/process/process.module';
     imports: [
         DatabaseModule,
         AuthModule,
-        MailModule,
+        MailTriggerModule,
         QueueModule,
         MicrosoftModule,
-        NestMailerModule,
+        MailModule,
         ProcessModule,
     ],
     providers: [
